+++ services/core/src/oci_cli_compute/compute_cli_extended.py
@@ -42,6 +42,8 @@ compute_cli.compute_root_group.add_command(compute_cli.boot_volume_attachment_gr
 compute_cli.compute_root_group.add_command(compute_cli.volume_attachment_group)
 
 # Disabling subclass commands
+compute_cli.volume_attachment_group.commands.pop(compute_cli.attach_volume_attach_service_determined_volume_details.name)
+compute_cli.volume_attachment_group.commands.pop(compute_cli.attach_volume_attach_emulated_volume_details.name)
 compute_cli.volume_attachment_group.commands.pop(compute_cli.attach_volume_attach_i_scsi_volume_details.name)
 
 compute_cli.compute_root_group.add_command(compute_cli.console_history_group)
@@ -93,7 +95,7 @@ or
 '{ "destinationUri": "https://objectstorage.us-phoenix-1.oraclecloud.com/n/MyNamespace/b/MyBucket/o/exported-image.qcow2", "destinationType": "objectStorageUri" }'"""
 cli_util.update_param_help(compute_cli.export_image, 'destination_type', "", append=True, example=destination_type_example)
 
-cli_util.get_param(compute_cli.attach_volume, 'type').type = custom_types.CliCaseInsensitiveChoice(['iscsi', 'paravirtualized'])
+cli_util.get_param(compute_cli.attach_volume, 'type').type = custom_types.CliCaseInsensitiveChoice(['service_determined', 'emulated', 'iscsi', 'paravirtualized'])
 
 cli_util.update_param_help(compute_cli.create_image, 'image_source_details', """[DEPRECATED] The use of the `oci compute image create` command to import an image from Object Storage is deprecated.