+++ services/file_storage/tests/extend_test_file_storage.py
@@ -1,0 +1,8 @@
+# coding: utf-8
+# Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
+
+# These mappings are used by generated tests to look up operations that have been moved in code in the CLI.
+MOVED_COMMANDS = {
+    ('fs', 'file_system', 'change-compartment'): ['fs', 'file-system', 'change-compartment'],
+    ('fs', 'mount_target', 'change-compartment'): ['fs', 'mount-target', 'change-compartment']
+}